WIP: Rename fcore to core
