+++ framework/wazuh/core/active_response.py
