+++ framework/wazuh/core/core_utils.py
